Fix ESLint errors in session.ts
The main problem is the circular dependency between `session.ts` and
`annotations.ts`. To fix it, split the second half of `session.ts` into a
separate file, like what is currently done with `project.ts` and
`project-implementation.ts`.
